+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This file is used to list changes made in each version of the nfs cookbook.
 
 - Sous Chefs adoption
 - Loosen version pin on line cookbook
+- Fix CentOS 8+ and Fedora and properly manage /etc/nfs.conf
 
 ## 4.0.0 - *2021-09-11*